requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
discord.py==1.3.1
dnspython==1.15.0
dnspython3==1.15.0
idna==2.8
mcstatus==2.3.0
multidict==4.7.4
requests==2.22.0
six==1.14.0
urllib3==1.25.8
websockets==8.1
yarl==1.4.2
